Enhancing Logo Image using some editing
I use Github in Dark mode ,  the logo of shuffleDNS is a png and blends with the dark background. Given below is a link of png with has outerglow , it will make it visible even in normal theme and dark theme. After accepting the pull request , replace the png with the one I have provided .
This is the png image link given below :
https://mega.nz/file/heBQHJiS#1-Y66gZtx33lZ5PV68AYXHJjzQIpwtz2WZO5WqdyaZQ